requirements.txt
# Azure SDK
azure-identity>=1.15.0
azure-cosmos>=4.5.0
azure-keyvault-secrets>=4.7.0

# Web framework
fastapi>=0.109.0
uvicorn>=0.25.0
gunicorn>=21.2.0
python-multipart>=0.0.6

# Authentication
msal>=1.26.0
PyJWT>=2.8.0

# Data handling
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Async/HTTP
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
